+++ tensorrt_llm/_torch/distributed/__init__.py
@@ -1,0 +1,20 @@
+from .communicator import Distributed, MPIDist, PPComm, TorchDist
+from .ops import (AllReduce, AllReduceFusionOp, AllReduceParams,
+                  AllReduceStrategy, DeepseekAllReduce, allgather, allreduce,
+                  reducescatter, userbuffers_allreduce_finalize)
+
+__all__ = [
+    "allgather",
+    "allreduce",
+    "reducescatter",
+    "userbuffers_allreduce_finalize",
+    "AllReduce",
+    "AllReduceParams",
+    "AllReduceFusionOp",
+    "AllReduceStrategy",
+    "DeepseekAllReduce",
+    "TorchDist",
+    "PPComm",
+    "MPIDist",
+    "Distributed",
+]
