+++ tensorrt_llm/_torch/distributed/communicator.py
@@ -1,13 +1,15 @@
+import atexit
 import os
 from abc import ABC, abstractmethod
+from typing import Optional
 
 import numpy as np
 import torch
 import torch.distributed as dist
 
-from ..._utils import (mpi_allgather, mpi_barrier, mpi_broadcast, mpi_comm,
-                       mpi_isend, mpi_recv)
-from ...mapping import Mapping
+from tensorrt_llm._utils import (mpi_allgather, mpi_barrier, mpi_broadcast,
+                                 mpi_comm, mpi_isend, mpi_recv)
+from tensorrt_llm.mapping import Mapping
 
 
 class Distributed(ABC):
@@ -192,3 +194,45 @@ class TorchDist(Distributed):
             self.broadcast_tp(obj, root)
         else:
             pass
+
+
+class PPComm:
+    # PP communication using torch.distributed with nccl backend
+    def __init__(self, global_mapping: Mapping):
+        self.mapping = global_mapping
+        if not dist.is_initialized():
+            master_ip = os.getenv("MASTER_ADDR", "localhost")
+            master_port = os.getenv("MASTER_PORT", "6000")
+            init_method = f"tcp://{master_ip}:{master_port}"
+            dist.init_process_group(backend="nccl",
+                                    init_method=init_method,
+                                    world_size=global_mapping.world_size,
+                                    rank=global_mapping.rank)
+            atexit.register(self._cleanup)
+
+        # Force NCCL initialization and rank population via PyTorch distributed barrier.
+        # This is necessary for NOW if using pp + tp because our custom nccl allreduce
+        # op for tp groups can interfere with PyTorch's NCCL initialization when PyTorch
+        # distributed performs the first comm. op and kick off nccl init. The barrier here
+        # ensures proper NCCL setup and GPU-procs binding at beginning.
+        dist.barrier(device_ids=[torch.cuda.current_device()])
+
+    def _cleanup(self):
+        if dist.is_initialized():
+            dist.destroy_process_group()
+
+    def send(self,
+             tensor: torch.Tensor,
+             dest: Optional[int] = None,
+             tag: Optional[int] = None):
+        if dest is None:
+            dest = self.mapping.next_pp_rank()
+        dist.send(tensor, dest, tag=tag)
+
+    def recv(self,
+             tensor: torch.Tensor,
+             src: Optional[int] = None,
+             tag: Optional[int] = None):
+        if src is None:
+            src = self.mapping.prev_pp_rank()
+        dist.recv(tensor, src, tag=tag)
